changed linelists to float32 in python and real to save memory
linelists are logged crosssections so no need for double/float64
precision

fixed this, should save more memory

+++ main_mod.f90
@@ -39,7 +39,7 @@ contains
     double precision,intent(inout) :: inwavenum(:)
     real,intent(inout) :: inlinetemps(:)
     real,intent(in) :: inpress(:)
-    double precision,intent(inout):: linelist(:,:,:,:)
+    real,intent(inout):: linelist(:,:,:,:)
     double precision,intent(inout):: bff(:,:)
     real, dimension(nciatemps), intent(in):: ciatemp
     real, intent(inout) :: cia(:,:,:)
